Move leader-key prefix subsystem from app_ui.go into app_prefix.go
app_ui.go (491 lines, near the 500-line guard) bundled three concerns:
pane focus navigation, the entire leader-key prefix command state machine
(~290 lines), and layout. The prefix subsystem is self-contained and its
rendering half already lives in app_prefix_palette.go, so the
command/dispatch half was split across the wrong file. Move the prefix
types, table, and methods (isPrefixKey..sendPrefixToTerminal) into a new
app_prefix.go; app_ui.go keeps focus navigation + layout and drops to 198
lines. Same-package move, zero renames or behavior change.
